autodetect correct kded touchpad module
plasma 5.24 renamed the "touchpad" kded module to "kded_touchpad".
fix this by listing all loaded modules in kded and look for the correct
module name to use.

Otherwise it would try to use the wrong one, fail to initialize,
and not re-enable the touchpad after resuming after suspend
leaving the user with an unusable touchpad and no easy way to enable it.

This now also gives an error message and exits, if no touchpad module is
loaded.

Fixes: #9